Pulled tile-preview Intro into a presenter
We're about to start adding at least one more similar intro, so to
prevent the logic in the controller and view for deciding just which
intros to show from getting convoluted, we introduce a presenter class.
This class is in charge of rendering the data intro.js needs on
elements, while allowing us to easily specify an ordering between them,
and enable or disable each independently.