Fixes for mass assignment and IDOR security holes
Overly-wide mass assignment by blindly creating objects off of params,
and insecure direct object references (IDOR) are two sides of the same
coin. Both are caused by being undercritical of input coming in from the
client. Mass assignment lets you write what you ought not be able to
write, and IDOR lets you read what you ought not be able to read. Both
tend to crop up in the same classes. As such, it's natural to look for
both kinds of holes at once.

In particular, in this commit we fix:

* Mass assignment in: AccountsController, BillingInformationsController,
  PrizesController, and SettingsController

* IDOR (usually having to do with looking at a tile in a board you're
  not actually in) in: ClientAdmin::TileCompletionsController,
  FriendshipsController, Invitation::AcceptancesController,
  and TileCompletionsController

In addition, we found some chunks of dead code in the process and
happily threw them away. That's always enjoyable.