Resolve invalid characters in subject line
Move from escaping invalid characters in links to fully removing them

The donwside of this approach is that mixpanel with not include special chars in the the pings and so subject lines on mixpanel will not reflect emoji usage etc.  Since this is currently a manual process, this will be fine.

In the future, we should rewrite the email process to include saaving emails to the db, which we can use to query specific email data to send to mixpanel as well as use to build out a client_admin self reporting module around email analytics.