Finish UI and begin translation from mock backend
The UI is ready to be transistioned from teh mock backend to the real backend that Herby has been working on.  Next steps are to figure out the logic to cleanly pass data to BoardStatsLineChartForm in order to plot the graph and retrive the data points for the activity_num and the header data points.  The methods "activity_num" and "action_type_class" are necessary in BoardStatusLineChartForm in order to maintain the current UI.

dsj