Remove es6 polyfills from application.js
The polyfills caused a bug in IE for client admins on tile creation.  This is more of a vulnerability in the IE implementation than in the polyfill implementation and should ultimately be resolved such that we can simply include polyfills across the entire app.

For now, I've only included es6 polyfills in the admin app.