Spring Cleaning Bug Sprint
_____

Closes #491 Update the scope for user search

When a user performs a user search to fink people to add,
it will now scope the search to users in the current board.
_____

Closes #478 Links in Explore Digest

Links in Explore Digest Emails route to Explore, but then
launch the specific tile preview based on the param "requested_tile_id".
_____

Closes #220

Users who invite spouses that are already in the board or invite themselves
will receive an error message letting them know the user is already in the board.
_____

Closes #283 Rewrites More Acts Buttons

Previous implemntation used lagacy rails/js patterns.
Rewrote how we retrieve and paginate acts in the activity feed
on /activity and /users/slug.
_____

Closes #476 Cleanup Mixpanel for this Sprint

Removes 'acted' pings, pings related to prize modals, 'viewed page' pings
(since we get those for free), and 'email sent' pings aross app. This was
done to lower Mixpanel costs. We can now move from legacy pricing to the
Business Plan, which is $999/yr, which will save us $3200/yr.
_____

Closes #227 Fix Activation date issue on tile drag

Small change to activation date z-index on tile thumbnails that fixes style bug.
_____

Closes #495 Fix styling for tiles on mobile

Fixes styling for users when viewing tiles on mobile. Adds correct margin for
tile prev/next buttons on mobile.  Adds missed custom colors for the custom
color palette on mobile. The way that we manage media quereies for this part
of the app is overly convoluted and prevents rapid dev for screen vs mobile.
This should be architected to enable more rapid design.
_____

Closes #224 Add Launch Date to Admin Demo Form

Adds launch date to admin demo form and restructure the form to look nicer.
_____

Closes #486 Fix Error Placement for Tile Image

Places the no image error message above the image menu in an incomplete tile.
Changes the message from the default message to an image specific message.
_____

Closes #484 Tooltips on tile stats

Regression bug caused from an erb to haml port.