Spring Cleaning Bug Sprint v2
Closes #503 Copy Board Regression and Sales Dashboard Issue

Copy board copies tiles to posted rather than draft, resolving a
regression that was created when we delegated the CopyBoard
copy tile function to the CopyTile service.

Sales dahboard can now handle sales orgs that have no users.
This introduces a potential use case where we need to be able to
handle orgs that don't have users.

_____

Closes #500 Explore tile creator with no creator

When no creator is present in an explore tile, the creator healine will
just reference the org.