Turn on User Search and move reindexing to backgnd
Closes #515 Turn on User search

User search is available to all users.

_____

Closes #508 Explore elastic search bugs

ES concurrency issues will be resolved by #193.
When tiles are not save in bulk removals, we will
not see the ES concurrency problems.  If this issue
comes up again, we should my to a queing system for
object reindexes, so we can bulk update ES records.

_____

Misc.

Adds model test coverage for Campaign.
Adds better reindex strategy to Campaigns
and Tiles.
Adds test coverage for User search.
Adds daily rake task for Eslatic Search updates,
specifically, a nightly reindex of campaigns.
_____

Closes #482 Simpler solution to merging orgs

Contracts can now be moved to whichever organization
from the contract edit view.
_____

Closes #520 Fix image loading for CA Explor single tile

Add js necessary to display image for explore single tile
preview (non-modal) fo CAs.

_____

Closes #518 UI issues with like to Sug Box from Email

The link to the suggesstion box that a CA gets when a
user suggests a tile will route them to the suggestion box
witht he correct UI activated.  If the CA is not in the board
that the link is linking to, they will be routes to Explore.
This scenario is only for SAs.