Closes #1024 Adds Tile Targeted Message Feature
All the JS related to this feature is nested on
/javascripts/client_admin/tile_stats.  The entire feature is managed on
the front end using a combination of JSON apis and Handlebars templates.

Adds alterations to the orginal backend API of Tile Targeted Messaging.
This includes small alterations to which users are targeted and how we
identify Tile answers.

Adds a js flash utility that easily allows flash messages within smaller
components (sub-component-flash.js).

Adds quill.js as the new wysiwyg editor in Airbo, with plans to use it
in TileBuilder.  This includes some customization to the editor to
include a custom 'Variable' format, which facilitates the variable embed
feature in Tile Targeted Messaging. This customization was written in
ES6 then transpiled.  The original ES6 is commented at the top of
quill/variable-embed.js. A standalone quill editor utility is found at
quill-standalone-editor.js.