Closes #1270 Resolves tickets loading bug
The issues was that points incrementing on the user was dependent on
Acts.  When I moved acts to a background job, points were not
incrementing in time for the ajax request to return the data to the dom.
 Acts should have nothing to do with user points, so we simply increment
the points after the tile completion is created.

Refactors and optimizes performance of PointIncrementer